server/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.6
numpy==1.14.5
opencv-python==3.4.1.15
pytesseract==0.2.4
PyYAML==3.13
requests==2.19.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
pymongo==3.7.2